+++ templates/create_table.sql
@@ -1,4 +1,4 @@
-CREATE TABLE public.model
+CREATE TABLE public.{{ table }}
 (
     {% for column in columns %}
     {{ column }}{% if not loop.last %},{% endif %}