+++ tests/test_table.py
@@ -5,7 +5,7 @@ from attrs_to_sql import attrs_to_table
 
 
 @attr.s(auto_attribs=True)
-class Model:
+class SampleModel:
     id: int = attr.ib(metadata={"primary_key": True, "type": "bigint", "auto_inc": True})
     title: str = attr.ib(metadata={"not_null": True, "length": 30})
     ids: list = attr.ib(metadata={"type": "bigint[]"})
@@ -19,6 +19,6 @@ def test_attrs_to_table():
     with open("tests/data/model.sql", encoding="utf-8") as f:
         expected_sql = f.read()
 
-    actual_sql = attrs_to_table(Model)
+    actual_sql = attrs_to_table(SampleModel)
 
     assert actual_sql == expected_sql